Components: Extract the discussion panel as reusable components
